Feature/debug blog function (#65)
* update-blog bug fix

* fix bugs for api

* delete print

* change src to directory's URI

* add check for range

* modify code style
